Backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
motor==3.7.0
pymongo==4.10.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File uploads & validation
python-multipart==0.0.20
Pillow==11.1.0

# HTTP client (for external APIs)
httpx==0.28.1

# AI / ML
google-generativeai==0.8.4

# Image hosting
cloudinary==1.42.1

# Scheduling (weather cron job)
apscheduler==3.11.0

# Config
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
